#31111 とりあえず一通り使えるようにした。同時にコードの整理。 #31111 ログメッセージの一部変更。 #35943 自己ベストが渡されなくなっていたので修正。

+++ DTXManiaプロジェクト/コード/スコア、曲/CScoreIni.cs
@@ -1488,10 +1488,10 @@ namespace DTXMania
             if (nTotal == 0)
                 return 0.0;
 
-            int nAuto = nTotal - (nPerfect + nGreat + nGood + nPoor + nMiss);
-            double dbPERFECT率 = (100.0 * nPerfect / nTotal);
-            double dbGREAT率 = (100.0 * nGreat / nTotal);
-            double dbCOMBO率 = (100.0 * nCombo / (nTotal));
+            double nAuto = nTotal - ( nPerfect + nGreat + nGood + nPoor + nMiss );
+            double dbPERFECT率 = ( 100.0 * nPerfect / nTotal );
+            double dbGREAT率 = ( 100.0 * nGreat / nTotal );
+            double dbCOMBO率 = ( 100.0 * nCombo / ( nTotal ) );
 
             if (nTotal == nAuto)
             {
